clean up when catching top-level exceptions
This is something that mmake did, but we lost in the
transition to a modern Error::Die (and it really must be
done, since otherwise you leave cruft that hurts other mfm
runs).

We add an array of cleanup functions to Error::Die to let
other users handle the same situation cleanly.