squash aliases in error reporting
If a rule 'foo.do' aliases to 'bar.do' and 'bar.do' fails,
the user ends up with:

  mfm: fatal: whatever
    ...while making foo (rule: bar.do)
    ...while making foo (rule: foo.do)

Since the use of 'bar' here is really an implementation
detail of 'foo', it doesn't make much sense to show it to
the user in the call stack.

We still show it if MFM_DEBUG is set, because it may be an
important part of debugging the foo and bar rules.