add 'replacesh' template rule
It is quite common to make a shell script by building up a
sequence of echo commands and cat'ing files. The quoting can
get quite messy, however. Instead, let's provide a tool that
lets callers specify an input template file and do simple
text replacements.

The format of the template file is

    input.sh
    VARIABLE1 source-file1
    VARIABLE2 source-file2
    ...

We always read source files with "head -n 1", which matches
the conf-* format.