add latex base rules
These rules build dvi from tex, and pdf and ps from dvi.
Direct building from tex to pdf is not supported by these
rules, but should be easy to add as a latex/pdfdirect module.

The rules are more or less a straight port from mmake. The
big changes are:

  - instead of constructing scripts manually, they are built
    from templates scripts using replacesh (which didn't
    exist in mmake)

  - the texdep mess is not necessary in mfm; we can just add
    the texdep attribute directly to the tex targets