mfm: ignore useless makefile targets
Files that are 'extragen' are marked as visible so that the
ycan be included in clean and ignore lists. Until now, the
visible marking also made mention of them in the Makefile.
However, this led to useless entries like:

  foo.aux:

  bar.aux:

that just clutter the Makefile. If a target is visible, but
would not contribute anything to the makefile (i.e., no
dependencies or commands), then let's just omit it.