reorganize asciidoc rules
The latex rules are split into a 'base', which handles
actual building, and a 'simple' which is good for quickly
building a document. The asciidoc rules, on the other hand,
are somewhat jumbled together. Let's split the asciidoc
rules in the same way.

For now, we'll just use a custom doc rule to build the MFM
docs.

While we're at it, let's use the nicer replacesh template
for building mkasciidoc (which didn't exist when it was
originally written).