asciidoc: add footer-suppressing asciidoc.conf
The "last updated" footer is annoying, and this seems to be
the only way to sanely suppress it ("-s" also suppresses the
header, which contains the nice style bits).

Users can still override the config by providing their own
asciidoc.conf, but it must have that name to match the
dependencies correctly.