mass renaming of mfm files
We have typically used the convention that input files to
mfm and intermediate files created by mfm are named using
all capital letters, and no other insignia.

On a case-insensitive filesystem, this can lead to problems
where the Makefile refers to "clean" or "all" as rules, but
the CLEAN and ALL files exist. Additionally, for mfm itself,
we have "mfm" the script and "MFM" the rules file.

Instead, let's prepend "MFM-" to each filename. We'll keep
the capital letters to make them stand out (and because
"MFM-clean" would look stupid).

The one exception is "MFM" itself, which has become
"MFM-RULES" (because "MFM-MFM" is stupid).