be more liberal in finding borrow paths
Given a path like foo/bar/baz, we originally checked for
"foo/bar/baz" in all rule paths. Later, we added checking
for the basename, "baz", in all rule paths (if we didn't
find the full path). But we never looked for "bar/baz".

This rule generalizes the lookup to searching for all sets
of path components, with longer preferred to shorter.

This is useful for borrowing sources into a subdirectory,
like finding "rules/perl/error/Error/Die.pm" when looking to
create "lib/Error/Die.pm".