place all perl modules into lib/
It is common to have perl module directories which differ
from a target name only in their case. For example, mfm
contains "MFM" as a directory, and "mfm" as a build target.
This can cause problems on case-insensitive filesystems.

This commit teaches the perldeps finder to look for files in
lib/ instead of in the current path. The mkperl script then
strips the lib/ when naming the file as part of the bundle.
This means that:

  1. All of your auto-depended perl modules need to be in lib/.

  2. Borrowed perl modules (thanks to the last commit's more
     liberal borrowing semantics) will end up in lib/.

To match this, we also move the MFM libraries into the lib/
directory.