make perl lib prefix configurable
The goal of the lib prefix is to prevent conflicts between
scripts and modules on case-insensitive systems. It comes at
the cost of having to put all of your libraries into lib/.
However, only some modules need this, and the others should
not have to pay the cost.

Strictly speaking, an option to simply turn off this feature
would have been sufficient, but you can configure it to an
arbitrary prefix by putting a string into MFM-PERLLIB. It
defaults to empty.