perl: use /usr/bin/env in stock conf-perl
This is more portable to systems with funny perl locations.

It is a little bit odd, because we use conf-perl for two
reasons: in the #!-line, and for testing the built script.
Which means we are actually running "/usr/bin/env perl -c"
to test. Wasteful, perhaps, but I don't think it's a problem
(and in a sense, good, because it tests that /usr/bin/env
actually works during the compilation phase).