use stemming in rule lookup
Since the mmake days, we have looked up borrowed files by
checking not only their full path, but also their basename.
In mfm, we take this even further, and for a path like
"1/2/3" we will look for "1/2/3", then "2/3", then "3".

We never did this stemming for rule lookups. It generally
didn't matter, as only $x.do and $x=* rule lookups will have
full paths at all. However, it can happen, and is made even
more annoying by the variable MFM-PERLLIB. We can't just put
$x.do into lib/$x.do, because we don't know what the perllib
prefix will be.

This patch implements borrow-style stemming for rule
lookups, which solves the problem. The stemming code is
adapted from the code in borrow(), but it must also handle
simultaneous multi-path lookup. So given a set of paths to
lookup like (1/2/x.do, 1/2/x=*), we use the order:

  1/2/x.do
  1/2/x=*
  2/x.do
  2/x=*
  x.do
  x=*

So that full paths are preferred over stemmed ones.