fix out-of-order stemming path lookup
The stemming in d1b694a was broken in two ways. In the
components subroutine, given two paths "sub/file.do" and
"sub/file=*", we should see:

  1. longer pathnames come before short (i.e., sub/file.do
     before file.do)

  2. the first path before the second (i.e., sub/file.do
     before sub/file=*)

Both were being violated. The first because we pop'd from
the end of the @expanded array (instead of shifting), and
the second because we must shift off the left of the
subarray.