improve stemming lookup rules
During a rule or borrow lookup, we search in each element of
the rule or borrow path, and then for each element, use
stemming to look for both the full path and any
right-anchored subset. So for a search path (a,b) a lookup
for "subdir/file" will check:

    a/subdir/file
    a/file
    b/subdir/file
    b/file

This is counter-intuitive, as a stray "file" in "a" can
override the full path in b. We should prefer full paths to
stemmed paths always, which can be accomplished by inverting
the nested loops.