dist: fix file::copy type problems
When doing "mfm dist", we feed the @FILES list to
File::Copy. However, that list is a mix of string filenames
and actual MFM::Target objects. Because MFM::Target
stringifies to the filename, in most cases this Just Works.
However, older versions of File::Copy get confused and die.

Let's just stringify the names manually, which takes care of
this problem. Arguably the @FILES list should be _just_
filenames in the first place, but it is very handy to do
things like:

  push @FILES, $TARGET;