clean: add MFM-TARGETS to FILES list
We depend on this for running "make clean", so we want to
make sure it gets shipped in "mfm dist". We put the change in
clean.do instead of in MFM-TARGETS.do because not all users
of MFM-TARGETS will necessarily want to ship it (only those
that ship a rule that looks at it).