run finalize code in a more sane environment
We simply ran the finalize code in the regular mfm
environment. This worked fine for the single user, which
just wrote out the MFM-TARGETS file. However, it is
reasonable for a rule to want to call something like
"formake" in the finalize subroutine. Previously this would
fail because $TARGET was not set properly. This patch sets
$TARGET properly so that the finalize subroutine runs with
it set, meaning it can use regular MFM calls.

Furthermore, we didn't wrap exceptions at all, so they
appeared to be root mfm errors. We should at least wrap them
in an Error::Parent to mention which rule triggered the
problem.