c: fix broken .lib handling in rule=x
This was a mis-port from the mmake rules. I accidentally
converted a "`cat $_`" into cat($_), thinking I was avoiding
a needless shell invocation. But of course this is perl, not
shell, so that shell invocation was not happening now, but
rather was being written into the Makefile.

Which is what obviously needs to happen, since reading the
contents of the lib file is only useful when you're actually
linking.