try all possible extensions when looking for rules
We previously just tried the final extension, so something
like "foo.bar.baz.bleep" would find only "rule.bleep.do".
Instead, we should try each of:

  rule.bar.baz.bleep.do
  rule.baz.bleep.do
  rule.bleep.do

I'm still unsure of the precedence ordering (which is
currently as shown above). Matching the longer one seems the
most sensible, though it is vulnerable to false positives if
an intended base has dots in it.