rules: refactor rule=replacesh into a generic form
This is a generally useful function for replacing tokens in
other languages (like perl), or data files. Let's pull it
out into its own rule, rule=replace.  We can then implement
replacesh in terms of replace.

We also add one minor feature: the input file line in the
ruledata can now specify multiple files. This is necessary
for rule=replacesh to be able to be implemented no top (it
needs to include warn-auto.sh, too).