simpleio: actually cat files in cat_scalar
We correctly slurped entire files, but we would quit after
hitting eof on the first file, meaning we didn't concatenate
anything. This never came up because I tend to use cat on a
single file, but the name and the code clearly imply that it
should be handling multiple files.