simpleio: read scalar refs literally
We already handle a scalar being passed (and treat it as a
filename), or a literal filehandle being passed. This adds
scalar refs, which are dereferenced and their contents read
literally.

You could also do this already with IO::String, but for such
a simple action, it's nice not to worry about the
dependency.