rules: add gitversion package
This just creates the MFM-VERSION file based on the output
of git-describe. It reads the package name from MFM-PACKAGE;
arguably these should be totally separate, and MFM-VERSION
should have _just_ the version. And I'm tempted to switch
that, but it's kind of nice to just have a single file that
ships in a dist tarball, which pollutes the directory a
little less.

We build the version file while mfm is run, rather than in
the Makefile. I was tempted to do it later, but:

  1. It's nice to have it available immediately for
     "mfm dist".

  2. We don't have a proper dependency for it, anyway,
     so are choices are to build every time, or never build
     after the first time. The latter is unacceptably
     infrequent. The former would work, but we would have to
     do a lot of magic not to cascade the update when it
     hasn't actually changed.