bootstrap: fix breakage with $VERSION
Bootstrapping has been broken since 45a7096, when we
introduced the use of the version.plib file. This file was
not included in the bootstrap run, causing perl to
rightfully complain about undefined variables.

Actually building and including it for the bootstrap is sort
of a chicken-and-egg problem. The simplest thing is to just
provide a "fake" version file for running the bootstrap.