rules/shell/install: be more flexible with basenames
When the hier function asks us to do:

  c /dest/path foo/bar

we always try to write to /dest/path/foo/bar. This may not
make sense if the "foo" part of the source path is some
internal organization in the project, and we may want to
write "dest/path/bar".

Changing "c" to require the full path would not be backwards
compatible. Therefore, we introduce a new "cb" (for
"c-basename") function that does expect the full path. And
it is trivial to implement "c" in terms of it.