rules/install: handle multiple hier functions
You can now build a hier.shlib file that has multiple hier
functions, and trigger only a subset of them by running
"do_install foo" (which will run hier_foo).

This allows you to use the same script for things like
manpage installation (which should not be required for
installing the rest of the package).