consistently order output of MFM::targets
We tried to sort on "$a cmp $b", but for objects that seems
to give semi-random ordering (even though stringification is
overloaded, it seems to use the perl pointer value).

This mean that finalize() calls were run in a random order,
meaning that gitignore would sometimes run before
MFM-TARGETS, and miss that it should be part of @CLEAN.

This should at least give us a consistent sort (and we do
not need to bother with priorities for this particular bug
because the names happen to sort as we want...yikes). It
also means we can drop the ad-hoc sorting in the Makefile
output.